requirements.txt
Django==4.2.7
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-extensions==3.2.3
djangorestframework